# docs/swagger.yaml
basePath: /
definitions:
...
  web.SusunanTimDetailResponse:
    properties:
      id:
        type: integer
      is_active:
        type: boolean
      keterangan:
        type: string
      level_jabatan:
        type: integer
      nama_jabatan:
        type: string
      nip:
        type: string
    type: object
...
  web.TimKerjaCreateRequest:
    properties:
      is_active:
        type: boolean
      keterangan:
        type: string
      kode_tim:
        type: string
      nama_tim:
        type: string
      tahun:
        type: string
    type: object
  web.TimKerjaDetailResponse:
    properties:
      id:
        type: integer
      kode_tim:
        type: string
      nama_tim:
        type: string
      susunan_tims:
        items:
          $ref: '#/definitions/web.SusunanTimDetailResponse'
        type: array
    type: object
  web.TimKerjaResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      keterangan:
        type: string
      kode_tim:
        type: string
      nama_tim:
        type: string
      tahun:
        type: string
      updated_at:
        type: string
    type: object
  web.TimKerjaUpdateRequest:
    properties:
      id:
        type: integer
      is_active:
        type: boolean
      keterangan:
        type: string
      kode_tim:
        type: string
      nama_tim:
        type: string
      tahun:
        type: string
    type: object
  web.WebResponse:
    description: Web Response
    properties:
      code:
        type: integer
      data: {}
      status:
        type: string
    type: object
host: ${HOST}:${PORT}
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API For Tim Kerja Services
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Tim Kerja Service API
  version: "1.0"
paths:
...
  /only_timkerja:
    get:
      consumes:
      - application/json
      description: Get All Tim Kerja
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.TimKerjaResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Get All Tim Kerja
      tags:
      - Tim Kerja
...
  /timkerja:
    get:
      consumes:
      - application/json
      description: Get All Tim Kerja with their Susunan Tim details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.TimKerjaDetailResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Get All Tim Kerja with Details
      tags:
      - Tim Kerja
    post:
      consumes:
      - application/json
      description: Create new Tim Kerja
      parameters:
      - description: Tim Kerja Create Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.TimKerjaCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.TimKerjaResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Create Tim Kerja
      tags:
      - Tim Kerja
  /timkerja/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Tim Kerja
      parameters:
      - description: Tim Kerja ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Delete Tim Kerja
      tags:
      - Tim Kerja
    get:
      consumes:
      - application/json
      description: Get Tim Kerja detail by ID
      parameters:
      - description: Tim Kerja ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.TimKerjaResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Get Tim Kerja by ID
      tags:
      - Tim Kerja
    put:
      consumes:
      - application/json
      description: Update Tim Kerja
      parameters:
      - description: Tim Kerja Update Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.TimKerjaUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.TimKerjaResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Update Tim Kerja
      tags:
      - Tim Kerja
swagger: "2.0"